решение Excercize 10 - Spring CRUD

+++ src/main/java/com/example/demo/dao/CountryRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface CountryRepository extends JpaRepository<Country, Long> {
+public interface CountryRepository extends JpaRepository<Country, Integer> {
     default List<Country> getCountryList(){
         return findAll();
     };
